+++ features/chat/messageOnPlayerDeath.js
@@ -1,24 +1,22 @@
 import settings from "../../settings";
 import * as triggers from '../../constants/triggers';
 import { getPlayerDeathMessage } from '../../utils/common';
-import { isInSkyblock } from '../../utils/playerState';
+import { getWorldName, isInSkyblock } from '../../utils/playerState';
+import { registerIf } from "../../utils/registers";
+import { CRIMSON_ISLE } from "../../constants/areas";
 
 const chatCommand = 'pc';
 
 triggers.KILLED_BY_TRIGGERS.forEach(entry => {
-    register(
-        "Chat",
-        (event) => {
-            sendMessageOnPlayerDeath({
-                isEnabled: settings.messageOnDeath
-            });
-        }
-    ).setCriteria(entry.trigger).setContains();
+	registerIf(
+		register("Chat", (event) => sendMessageOnPlayerDeath()).setCriteria(entry.trigger).setContains(),
+		() => settings.messageOnDeath && isInSkyblock() && getWorldName() === CRIMSON_ISLE
+	);
 });
 
-function sendMessageOnPlayerDeath(options) {
+function sendMessageOnPlayerDeath() {
 	try {
-		if (!options.isEnabled || !isInSkyblock()) {
+		if (!settings.messageOnDeath || !isInSkyblock() || getWorldName() !== CRIMSON_ISLE) {
 			return;
 		}